pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "kabot"
version = "0.1.0"
description = "Bitcoin prediction-market trading research system for Kalshi."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "python-dotenv>=1.0.1",
  "pandas>=2.2.0",
  "numpy>=2.0.0",
  "requests>=2.32.0",
  "psycopg[binary]>=3.2.0",
  "cryptography>=45.0.0",
  "websocket-client>=1.7.0",
  "tomli>=2.0.1; python_version < '3.11'",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
]

[project.scripts]
kabot = "kabot.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
